Version Update & Movement
# Description
Updated the Unity version due to issues with job leak warnings on the old version and the security concerns.

Modified the `TransformTweeningComponent` to instead reused the cache tween when still active and generate a new tween if completed or killed to create variation on the shake instead of reusing the same shake.

Base movement component was added.

Implemented a simple movement to target component for the objects to move towards the centre of the screen.

Included an event for when the interaction is ended and updated the interface accordingly.

Added edge particles to the bottom and top of the screen.