Player Controller & Animator
# Description
Added a `PlayerController` class to handle the player logic. Handles the player accelerating while holding the interact button and starts to decelerate once the button is let go.

Added component (`AnimatorModifierComponent`) that modifies an animator parameters using the hashed value contained in the `AnimatorParameterHasher` or applies the parameter changes from a `AnimatorParameterEvent`.

Added scriptable object `AnimatorParameterHasher` that on enable caches the hash value of the parameter name used by animators. Additionally it can verify if an animator contains the defined parameter. Taken from my `Wild_MajorJam7` repository and slightly modified.

Added abstract scriptable object `AnimatorParameterEvent` that is intended to on application change the value of the provided animator parameter. The following classes were added inheriting from it:
    - `AnimatorBoolParameterEvent`: Changes the animator parameter to the set bool value on application.
    -`AnimatorTriggerParameterEvent`: Triggers the animator trigger parameter on application.

Unity Events in `InputBinder` updated to use the player controller and the animator modifier component instead. In the case of the animator, it applies the animator parameter events to activate the corresponding animations.