Project Structure & Custom Package
# Description
Decided to change the project structure to contain the project inside a `_Project` folder to create a clear distinction between internal assets and third-party assets.

Remove the majority of the scripts inside the "Scripts/Animations" folder as they were transferred to a custom package that is now being included in the project. The package can be found in the following repository URL: https://github.com/Janfra/Janito-Unity-Packages.git