Projectile Rotation & Testing
# Description
Projectiles now face the direction they are moving to and are spawned facing the aimed direction.

Added `RotateComponent`  to rotate transform to align with the direction the rigidbody is moving to. Component may be renamed.

Adjusted the sorting order of all components inside the projectile to avoid projectiles appearing on top of middle hexagon or entities.

Scene was modified to more easily test projectiles and shooting.